+++ src/main/java/com/example/nukebombsimulation/calculations/GroundBomb.java
@@ -10,7 +10,7 @@ public class GroundBomb implements IRadiusCalculator
     }
 
     @Override
-    public int calculateRadius() {
-        return (int) ((9.8*Math.pow(energy,0.333))/(Math.pow(1 + Math.pow((3175 / energy), 2), 1/6)));
+    public double calculateRadius() {
+        return ((9.8*Math.pow(energy,0.333))/(Math.pow(1 + Math.pow((3175 / energy), 2), 1/6)));
     }
 }
