+++ src/main/java/com/example/nukebombsimulation/calculations/IRadiusCalculator.java
@@ -2,5 +2,5 @@ package com.example.nukebombsimulation.calculations;
 
 public interface IRadiusCalculator {
 
-    int calculateRadius();
+    double calculateRadius();
 }
